cleaned the code with new GPSData constructors
