+++ src/subframe_decoding.jl
@@ -1,59 +1,7 @@
+
 GPS_PI = 3.1415926535898
 
-struct TLM_HOW_Data_Struct
-    integrity_status_flag::Bool
-    TOW::Int64
-    alert_flag::Bool
-    anti_spoof_flag::Bool
-end
-
-
-struct Subframe_1_Data
-    trans_week::Int64
-    codeonl2::Int64
-    ura::Float64
-    svhealth::String
-    IODC::String
-    l2pcode::Bool
-    T_GD::Float64
-    t_oc::Int64
-    a_f2::Float64
-    a_f1::Float64
-    a_f0::Float64
-end
-
-struct Subframe_2_Data
-    IODE::String
-    C_rs::Float64
-    Δn::Float64
-    M_0::Float64
-    C_uc::Float64
-    e::Float64
-    C_us::Float64
-    sqrt_A::Float64
-    t_oe::Int64
-    fit_interval::Bool
-    AODO::Int64
-end
-
-struct Subframe_3_Data
-    C_ic::Float64
-    Ω_0::Float64
-    C_is::Float64
-    i_0::Float64
-    C_rc::Float64
-    ω::Float64
-    Ω_dot::Float64
-    IODE::String
-    IDOT::Float64
-end
-
-struct Subframe_4_Data
-end
-
-struct Subframe_5_Data
-end
-
+#include("gpsl1_structs.jl")
 
 """
     Decodes words of the first two Words of a subframe
@@ -335,248 +283,4 @@ function decode_subframe_5(words)
 
     TLM_HOW = decode_TLM_HOW(words[1:2])
     return TLM_HOW
-end
-
-function create_data(
-    TLM_HOW_Data::TLM_HOW_Data_Struct,
-    subfr_1_data::Subframe_1_Data,
-    subfr_2_data::Subframe_2_Data,
-    subfr_3_data::Subframe_3_Data)
-    
-
-    data = GPSData(
-        TLM_HOW_Data.integrity_status_flag,
-        TLM_HOW_Data.TOW,
-        TLM_HOW_Data.alert_flag,
-        TLM_HOW_Data.anti_spoof_flag,
-
-        subfr_1_data.trans_week,
-        subfr_1_data.codeonl2,
-        subfr_1_data.ura,
-        subfr_1_data.svhealth,
-        subfr_1_data.IODC,
-        subfr_1_data.l2pcode,
-        subfr_1_data.T_GD,
-        subfr_1_data.t_oc,
-        subfr_1_data.a_f2,
-        subfr_1_data.a_f1,
-        subfr_1_data.a_f0,
-
-        subfr_2_data.IODE,
-        subfr_2_data.C_rs,
-        subfr_2_data.Δn,
-        subfr_2_data.M_0,
-        subfr_2_data.C_uc,
-        subfr_2_data.e,
-        subfr_2_data.C_us,
-        subfr_2_data.sqrt_A,
-        subfr_2_data.t_oe,
-        subfr_2_data.fit_interval,
-        subfr_2_data.AODO,
-
-        subfr_3_data.C_ic,
-        subfr_3_data.Ω_0,
-        subfr_3_data.C_is,
-        subfr_3_data.i_0,
-        subfr_3_data.C_rc,
-        subfr_3_data.ω,
-        subfr_3_data.Ω_dot,
-        subfr_3_data.IODE,
-        subfr_3_data.IDOT
-    )
-    return data
-end
-
-
-function create_data_sub1(
-    TLM_HOW_Data::TLM_HOW_Data_Struct,
-    subfr_1_data::Subframe_1_Data)
-    
-
-    data = GPSData(
-        TLM_HOW_Data.integrity_status_flag,
-        TLM_HOW_Data.TOW,
-        TLM_HOW_Data.alert_flag,
-        TLM_HOW_Data.anti_spoof_flag,
-
-        subfr_1_data.trans_week,
-        subfr_1_data.codeonl2,
-        subfr_1_data.ura,
-        subfr_1_data.svhealth,
-        subfr_1_data.IODC,
-        subfr_1_data.l2pcode,
-        subfr_1_data.T_GD,
-        subfr_1_data.t_oc,
-        subfr_1_data.a_f2,
-        subfr_1_data.a_f1,
-        subfr_1_data.a_f0,
-
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing
-    )
-    return data
-end
-
-function create_data_sub2(
-    prev_data::GPSData,
-    TLM_HOW_Data::TLM_HOW_Data_Struct,
-    subfr_2_data::Subframe_2_Data)
-    
-
-    data = GPSData(
-        TLM_HOW_Data.integrity_status_flag,
-        TLM_HOW_Data.TOW,
-        TLM_HOW_Data.alert_flag,
-        TLM_HOW_Data.anti_spoof_flag,
-
-        prev_data.trans_week,
-        prev_data.codeonl2,
-        prev_data.ura,
-        prev_data.svhealth,
-        prev_data.IODC,
-        prev_data.l2pcode,
-        prev_data.T_GD,
-        prev_data.t_oc,
-        prev_data.a_f2,
-        prev_data.a_f1,
-        prev_data.a_f0,
-
-        subfr_2_data.IODE,
-        subfr_2_data.C_rs,
-        subfr_2_data.Δn,
-        subfr_2_data.M_0,
-        subfr_2_data.C_uc,
-        subfr_2_data.e,
-        subfr_2_data.C_us,
-        subfr_2_data.sqrt_A,
-        subfr_2_data.t_oe,
-        subfr_2_data.fit_interval,
-        subfr_2_data.AODO,
-
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing,
-        nothing
-    )
-    return data
-end
-
-function create_data_sub3(
-    prev_data::GPSData,
-    TLM_HOW_Data::TLM_HOW_Data_Struct,
-    subfr_3_data::Subframe_3_Data)
-    
-
-    data = GPSData(
-        TLM_HOW_Data.integrity_status_flag,
-        TLM_HOW_Data.TOW,
-        TLM_HOW_Data.alert_flag,
-        TLM_HOW_Data.anti_spoof_flag,
-
-        prev_data.trans_week,
-        prev_data.codeonl2,
-        prev_data.ura,
-        prev_data.svhealth,
-        prev_data.IODC,
-        prev_data.l2pcode,
-        prev_data.T_GD,
-        prev_data.t_oc,
-        prev_data.a_f2,
-        prev_data.a_f1,
-        prev_data.a_f0,
-
-        prev_data.IODE_Sub_2,
-        prev_data.C_rs,
-        prev_data.Δn,
-        prev_data.M_0,
-        prev_data.C_uc,
-        prev_data.e,
-        prev_data.C_us,
-        prev_data.sqrt_A,
-        prev_data.t_oe,
-        prev_data.fit_interval,
-        prev_data.AODO,
-
-        subfr_3_data.C_ic,
-        subfr_3_data.Ω_0,
-        subfr_3_data.C_is,
-        subfr_3_data.i_0,
-        subfr_3_data.C_rc,
-        subfr_3_data.ω,
-        subfr_3_data.Ω_dot,
-        subfr_3_data.IODE,
-        subfr_3_data.IDOT
-    )
-    return data
-end
-
-function create_data_new_TLM(
-    prev_data::GPSData,
-    TLM_HOW_Data::TLM_HOW_Data_Struct)
-    
-    data = GPSData(
-        TLM_HOW_Data.integrity_status_flag,
-        TLM_HOW_Data.TOW,
-        TLM_HOW_Data.alert_flag,
-        TLM_HOW_Data.anti_spoof_flag,
-
-        prev_data.trans_week,
-        prev_data.codeonl2,
-        prev_data.ura,
-        prev_data.svhealth,
-        prev_data.IODC,
-        prev_data.l2pcode,
-        prev_data.T_GD,
-        prev_data.t_oc,
-        prev_data.a_f2,
-        prev_data.a_f1,
-        prev_data.a_f0,
-
-        prev_data.IODE_Sub_2,
-        prev_data.C_rs,
-        prev_data.Δn,
-        prev_data.M_0,
-        prev_data.C_uc,
-        prev_data.e,
-        prev_data.C_us,
-        prev_data.sqrt_A,
-        prev_data.t_oe,
-        prev_data.fit_interval,
-        prev_data.AODO,
-
-        prev_data.C_ic,
-        prev_data.Ω_0,
-        prev_data.C_is,
-        prev_data.i_0,
-        prev_data.C_rc,
-        prev_data.ω,
-        prev_data.Ω_dot,
-        prev_data.IODE_Sub_3,
-        prev_data.IDOT
-    )
-    return data
 end
